backend/requirements.txt
django>=5.1
djangorestframework>=3.15
django-cors-headers>=4.3
langchain>=0.3
langchain-openai>=0.3
tiktoken>=0.7
python-dotenv>=1.0
